+++ Changes.md
@@ -6,6 +6,7 @@ API
 
 - SceneAlgo :
   - Added `parallelReduceLocations()` for implementing functions that need to combine results while traversing a ScenePlug.
+  - Added `hierarchyHash()` for hashing all children of a scene location.
 
 
 1.5.5.0 (relative to 1.5.4.1)